Export eslint rules as a function which make some common customisations simpler
