+++ packages/eslint-config-hmpps/CHANGELOG.md
@@ -1,5 +1,15 @@
 # History of changes
 
+## 0.0.1-alpha.6
+
+Pre-release for testing with HMPPS projects.
+eslint rules are now exported as a function which make some common customisations simpler.
+Migration has been made simpler for existing projects using:
+
+```shell
+npx @ministryofjustice/eslint-config-hmpps
+```
+
 ## 0.0.1-alpha.5
 
 Initial pre-release for testing with HMPPS projects.
