+++ packages/monitoring/eslint.config.mjs
@@ -1,4 +1,4 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import hmppsConfig from '@ministryofjustice/eslint-config-hmpps'
 
-export default hmppsConfig
+export default hmppsConfig()
